Progress on Transaction creation with Requirements.
First, Form was adapted to work better with arrays.
Previously, the manual DOM manipulation caused
elements to be created that the VDOM didn't know
about. A new technique has been adopted to
fix the problem.

SearchField has been updated to work inside a form
more readily. Previously, it was a problem child
and now it's just a bit truant.